+++ packages/cli/rsbuild-plugin-esbuild/src/index.ts
@@ -91,6 +91,8 @@ export function pluginEsbuild(
             .use(ESBuildMinifyPlugin)
             .init(
               () =>
+                // @ts-expect-error
+                // Due to rspack compiler missing some properties, but ESBuildMinifyPlugin only works in webpack.
                 new ESBuildMinifyPlugin({
                   // other legalComments such as linked is not supported yet
                   // https://github.com/privatenumber/esbuild-loader/issues/263