+++ packages/solutions/app-tools/src/builder/shared/builderPlugins/adapterSSR.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { isHtmlDisabled, RsbuildPlugin, BundlerChain } from '@rsbuild/shared';
+import { isHtmlDisabled, RsbuildPlugin, RspackChain } from '@rsbuild/shared';
 import { mergeRsbuildConfig } from '@rsbuild/core';
 import { fs, isUseSSRBundle } from '@modern-js/utils';
 import type { HtmlWebpackPlugin } from '@modern-js/uni-builder';
@@ -11,7 +11,7 @@ import type {
 } from '../../../types';
 import { HtmlAsyncChunkPlugin, RouterPlugin } from '../bundlerPlugins';
 import type { BuilderOptions } from '../types';
-import { getServerCombinedModueFile } from '../../../analyze/utils';
+import { getServerCombinedModueFile } from '../../../plugins/analyze/utils';
 
 export const builderPluginAdapterSSR = <B extends Bundler>(
   options: BuilderOptions<B>,
@@ -117,7 +117,7 @@ function applyAsyncChunkHtmlPlugin({
   modernConfig,
   HtmlBundlerPlugin,
 }: {
-  chain: BundlerChain;
+  chain: RspackChain;
   modernConfig: AppNormalizedConfig<'shared'>;
   HtmlBundlerPlugin: any;
 }) {
@@ -129,7 +129,7 @@ function applyAsyncChunkHtmlPlugin({
 }
 
 function applyRouterPlugin<B extends Bundler>(
-  chain: BundlerChain,
+  chain: RspackChain,
   pluginName: string,
   options: Readonly<BuilderOptions<B>>,
   HtmlBundlerPlugin: typeof HtmlWebpackPlugin,
@@ -170,7 +170,7 @@ function applyFilterEntriesBySSRConfig({
   appNormalizedConfig,
 }: {
   isProd: boolean;
-  chain: BundlerChain;
+  chain: RspackChain;
   appNormalizedConfig: AppNormalizedConfig<'shared'>;
 }) {
   const { server: serverConfig, output: outputConfig } = appNormalizedConfig;
@@ -222,7 +222,7 @@ function applyFilterEntriesBySSRConfig({
 }
 
 async function applySSRLoaderEntry<B extends Bundler>(
-  chain: BundlerChain,
+  chain: RspackChain,
   optinos: BuilderOptions<B>,
   isServer: boolean,
 ) {
@@ -261,7 +261,7 @@ async function applySSRLoaderEntry<B extends Bundler>(
 }
 
 function applySSRDataLoader<B extends Bundler>(
-  chain: BundlerChain,
+  chain: RspackChain,
   options: BuilderOptions<B>,
 ) {
   const { normalizedConfig, appContext } = options;