+++ packages/cli/plugin-ssg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-ssg
 
+## 2.53.0
+
+### Patch Changes
+
+- @modern-js/prod-server@2.53.0
+- @modern-js/utils@2.53.0
+
 ## 2.52.0
 
 ### Patch Changes